% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Scenario data}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\source{

}
\usage{
data(metadata)
}
\description{
Metadata for the parameters that are used in CEMRA
}
\examples{
data(metadata)
}
\references{

}
\keyword{datasets}
